lab-1: Added API-gateway and test for monolith

+++ monolith/src/main/java/ru/ifmo/monolith/dto/CityDto.java
@@ -10,5 +10,5 @@ import lombok.Data;
 public class CityDto {
 
     private String name;
-    private Long hotelsCount;
+    private Integer hotelsCount;
 }
